please dont mess up

+++ src/utility/Sorting.java
@@ -2,9 +2,9 @@ package utility;
 
 public class Sorting {
 
-    public static void bubbleSort(int[] arr) {
+    public static void bubbleSort(double[] arr) {
         int n = arr.length;
-        int temp = 0;
+        double temp = 0;
         int i = 0;
         int j = 0;
         for (i = 0; i < n; i++) {
@@ -20,7 +20,7 @@ public class Sorting {
         }
     }
 
-    public static void quickSort(int[] arr, int low, int high) {
+    public static void quickSort(double[] arr, int low, int high) {
         if (low < high) {
             int pi = partition(arr, low, high);
 
@@ -30,8 +30,8 @@ public class Sorting {
         }
     }
 
-    public static int partition(int[] arr, int low, int high) {
-        int pivot = arr[high];
+    public static int partition(double[] arr, int low, int high) {
+        double pivot = arr[high];
         int i = low - 1; // index of smaller element
 
         for (int j = low; j < high; j++) {
@@ -40,7 +40,7 @@ public class Sorting {
                 i++;
 
                 // swap arr[i] and arr[j]
-                int temp = arr[i];
+                double temp = arr[i];
                 arr[i] = arr[j];
                 arr[j] = temp;
 
@@ -48,15 +48,13 @@ public class Sorting {
         }
 
         // swap arr[i+1] and arr[high] (or pivot)
-        int temp = arr[i + 1];
+        double temp = arr[i + 1];
         arr[i + 1] = arr[high];
         arr[high] = temp;
 
         // return the index of the pivot
         return i + 1;
     }
-
-
 
     public static void selectionSort(double[] arr) {
         int n = arr.length;
@@ -72,8 +70,7 @@ public class Sorting {
             arr[i] = temp;
         }
     }
-    
-    
+
     public static void mergeSort(double[] array) {
         if (array == null || array.length <= 1) {
             return;
@@ -117,8 +114,4 @@ public class Sorting {
         }
     }
 
-    
 }
-
-
-
